pyproject.toml
[tool.poetry]
name = "tpmla"
version = "0.1.0"
description = "tp mla itba 2022"
authors = ["ivenini <ivenini@luminaamericas.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "2.2.0"
SQLAlchemy = "1.3.24"
apache-airflow-providers-postgres = "2.3.0"
matplotlib = "3.3.4"
scikit-learn = "0.24.2"
boto3 = "^1.21.26"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
